Scraper/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.0
lxml==5.1.0
fake-useragent==1.4.0
feedparser==6.0.12
tabulate==0.9.0
python-dotenv==1.0.0
selenium==4.27.1
webdriver-manager==4.0.2
